requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-admin-notification==0.2
django-admin-persian-fonts==0.2
django-ckeditor==6.5.0
django-jalali==6.0.0
django-js-asset==2.0.0
django-render-partial==0.4
django-rest-framework==0.1.0
django-social-share==2.3.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
idna==3.3
install==1.3.5
jdatetime==4.1.0
Pillow==9.2.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
